refactor Query component
