+++ plugin-hrm-form/src/translations/en-US/flexUI.json
@@ -76,6 +76,8 @@
   "SearchResults-Contacts": " Contacts",
   "SearchResults-Case": "1 Case",
   "SearchResults-Cases": " Cases",
+  "SearchResults-For": " for",
+
   
   "StandaloneSearch-SideNav": "Search",
   "CaseHeader-Opened": "Opened",
